requirements.txt
apache-airflow==2.10.4
praw==7.8.1
apache-airflow-providers-amazon==9.2.0
pandas==2.2.3
numpy==2.2.2
pendulum==3.0.0